Use u128 as the container for constant values, several sign extension bugs

+++ borealis/borealis/src/boom/mod.rs
@@ -291,7 +291,7 @@ pub enum Type {
         size: Size,
     },
 
-    Constant(i64),
+    Constant(i128),
 
     Union {
         name: InternedString,
