+++ brig/kernel/src/host/dbt/x86/encoder/instructions/and.rs
@@ -23,7 +23,7 @@ pub fn encode<A: Alloc>(assembler: &mut CodeAssembler, src: &Operand<A>, dst: &O
                 width_in_bits: Width::_8,
             },
         ) => {
-            if *left > i32::MAX as u64 {
+            if *left > i32::MAX as u128 {
                 panic!("AND immediate too large: {left:x}");
             }
             assembler
@@ -37,7 +37,7 @@ pub fn encode<A: Alloc>(assembler: &mut CodeAssembler, src: &Operand<A>, dst: &O
                 width_in_bits: Width::_16,
             },
         ) => {
-            if *left > i32::MAX as u64 {
+            if *left > i32::MAX as u128 {
                 panic!("AND immediate too large: {left:x}");
             }
             assembler
@@ -51,7 +51,7 @@ pub fn encode<A: Alloc>(assembler: &mut CodeAssembler, src: &Operand<A>, dst: &O
                 width_in_bits: Width::_32,
             },
         ) => {
-            if *left > i32::MAX as u64 {
+            if *left > i32::MAX as u128 {
                 panic!("AND immediate too large: {left:x}");
             }
             assembler
@@ -65,7 +65,7 @@ pub fn encode<A: Alloc>(assembler: &mut CodeAssembler, src: &Operand<A>, dst: &O
                 width_in_bits: Width::_64,
             },
         ) => {
-            if *left > i32::MAX as u64 {
+            if *left > i32::MAX as u128 {
                 panic!("AND immediate too large: {left:x}");
             }
             assembler
@@ -87,7 +87,7 @@ pub fn encode<A: Alloc>(assembler: &mut CodeAssembler, src: &Operand<A>, dst: &O
                 width_in_bits: Width::_64,
             },
         ) => {
-            if *left > i32::MAX as u64 {
+            if *left > i32::MAX as u128 {
                 panic!("AND immediate too large: {left:x}");
             }
             assembler
@@ -112,7 +112,7 @@ pub fn encode<A: Alloc>(assembler: &mut CodeAssembler, src: &Operand<A>, dst: &O
                 width_in_bits: Width::_8,
             },
         ) => {
-            if *left > u8::MAX as u64 {
+            if *left > u128::from(u8::MAX) {
                 panic!("AND immediate too large: {left:x}");
             }
             assembler
